+++ src/RcppExports.cpp
@@ -79,8 +79,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // mediation_analysis_cpp
-void mediation_analysis_cpp(NumericMatrix data, CharacterVector column_names, IntegerVector exposure_col_idx, IntegerVector mediator_col_idx, IntegerVector outcome_col_idx, int nrep, std::string output_file, Rcpp::Nullable<Rcpp::NumericVector> weights, std::string pert, uint64_t base_seed, std::string mediator_family, std::string outcome_family, bool replace_outcome, std::string output_format, double treat_value, double control_value, int chunk_size, int grain_size, bool overwrite, bool excel_safe_csv, bool match_mediation);
-RcppExport SEXP _fastmed_mediation_analysis_cpp(SEXP dataSEXP, SEXP column_namesSEXP, SEXP exposure_col_idxSEXP, SEXP mediator_col_idxSEXP, SEXP outcome_col_idxSEXP, SEXP nrepSEXP, SEXP output_fileSEXP, SEXP weightsSEXP, SEXP pertSEXP, SEXP base_seedSEXP, SEXP mediator_familySEXP, SEXP outcome_familySEXP, SEXP replace_outcomeSEXP, SEXP output_formatSEXP, SEXP treat_valueSEXP, SEXP control_valueSEXP, SEXP chunk_sizeSEXP, SEXP grain_sizeSEXP, SEXP overwriteSEXP, SEXP excel_safe_csvSEXP, SEXP match_mediationSEXP) {
+void mediation_analysis_cpp(NumericMatrix data, CharacterVector column_names, IntegerVector exposure_col_idx, IntegerVector mediator_col_idx, IntegerVector outcome_col_idx, int nrep, std::string output_file, Rcpp::Nullable<Rcpp::NumericVector> weights, std::string pert, uint64_t base_seed, std::string mediator_family, std::string outcome_family, bool replace_outcome, std::string output_format, double treat_value, double control_value, int chunk_size, int grain_size, bool overwrite, bool excel_safe_csv, bool match_mediation, std::string loop_order);
+RcppExport SEXP _fastmed_mediation_analysis_cpp(SEXP dataSEXP, SEXP column_namesSEXP, SEXP exposure_col_idxSEXP, SEXP mediator_col_idxSEXP, SEXP outcome_col_idxSEXP, SEXP nrepSEXP, SEXP output_fileSEXP, SEXP weightsSEXP, SEXP pertSEXP, SEXP base_seedSEXP, SEXP mediator_familySEXP, SEXP outcome_familySEXP, SEXP replace_outcomeSEXP, SEXP output_formatSEXP, SEXP treat_valueSEXP, SEXP control_valueSEXP, SEXP chunk_sizeSEXP, SEXP grain_sizeSEXP, SEXP overwriteSEXP, SEXP excel_safe_csvSEXP, SEXP match_mediationSEXP, SEXP loop_orderSEXP) {
 BEGIN_RCPP
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< NumericMatrix >::type data(dataSEXP);
@@ -104,7 +104,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< bool >::type overwrite(overwriteSEXP);
     Rcpp::traits::input_parameter< bool >::type excel_safe_csv(excel_safe_csvSEXP);
     Rcpp::traits::input_parameter< bool >::type match_mediation(match_mediationSEXP);
-    mediation_analysis_cpp(data, column_names, exposure_col_idx, mediator_col_idx, outcome_col_idx, nrep, output_file, weights, pert, base_seed, mediator_family, outcome_family, replace_outcome, output_format, treat_value, control_value, chunk_size, grain_size, overwrite, excel_safe_csv, match_mediation);
+    Rcpp::traits::input_parameter< std::string >::type loop_order(loop_orderSEXP);
+    mediation_analysis_cpp(data, column_names, exposure_col_idx, mediator_col_idx, outcome_col_idx, nrep, output_file, weights, pert, base_seed, mediator_family, outcome_family, replace_outcome, output_format, treat_value, control_value, chunk_size, grain_size, overwrite, excel_safe_csv, match_mediation, loop_order);
     return R_NilValue;
 END_RCPP
 }
@@ -115,7 +116,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_fastmed_fastmed_test_two_bootstrap_samples", (DL_FUNC) &_fastmed_fastmed_test_two_bootstrap_samples, 4},
     {"_fastmed_fastmed_test_ols_sigmas_cpp", (DL_FUNC) &_fastmed_fastmed_test_ols_sigmas_cpp, 3},
     {"_fastmed_glm_fit_cpp", (DL_FUNC) &_fastmed_glm_fit_cpp, 8},
-    {"_fastmed_mediation_analysis_cpp", (DL_FUNC) &_fastmed_mediation_analysis_cpp, 21},
+    {"_fastmed_mediation_analysis_cpp", (DL_FUNC) &_fastmed_mediation_analysis_cpp, 22},
     {NULL, NULL, 0}
 };